exercise 2.5 courseinfo step10 complete - separate out subcomponents
